+++ packages/server/templates/login.mustache
@@ -1,8 +1,5 @@
 {{> header}}
 <h1>Sign in</h1>
-{{#errors.query}}
-<div>{{msg}}</div>
-{{/errors.query}}
 <form action="#" method="POST">
   <div>
     <label for="email" class="block">Email</label>
